Revert "feat: add new non TLS grpc server for healthcheck"
This reverts commit 9184d0e5cf6562c258d585efa9490d045ab3bc6f.
